api/requirements.txt
Django==2.2.5
djangorestframework==3.11.0
ipdb==0.11
mysqlclient==1.3.13
django-cors-headers==3.0.2